index-log-tracking/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pytz
databricks-sql-connector
azure-identity
azure-keyvault-secrets


azure-search-documents==11.6.0b12
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0b1  # beta version required for newer APIs

# LangChain + Ecosystem
langchain>=0.1.14
langchain-openai>=0.1.3
langchain-community>=0.0.30

# OpenAI
openai>=1.0.0
tiktoken>=0.5.1  # for token counting

# Document parsing
PyPDF2>=3.0.1
openpyxl>=3.1.2
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.1.0

# For Jupyter environment support
ipykernel>=6.29.0